updated file - restructured and added logger

+++ src/make_voc_dataset/make_voc_dataset.py
@@ -6,6 +6,11 @@ import mmcv
 from tqdm import tqdm
 import shutil
 import numpy as np
+import logging
+
+logging.basicConfig(
+    format="[ %(asctime)s ] -- %(message)s", datefmt="%m/%d/%Y %I:%M:%S %p"
+)
 
 "Required image path and the xmls path and tha path to save the voc_structure"
 
@@ -22,9 +27,8 @@ def create_blank_voc_dir_str(dest_path):
 def make_voc_dataset(opt):
     file_list = list()
 
-    base_path = opt.base_path
+    base_path = opt.source_path
     dest_path = opt.dest_path
-    create_blank_voc_dir_str(dest_path)
 
     dest_annotation_path = os.path.join(
         dest_path, "VOCdevkit", "VOC2007", "Annotations"
@@ -38,38 +42,68 @@ def make_voc_dataset(opt):
         file_list += [os.path.join(dirpath, file) for file in filenames]
 
     images = [x for x in file_list if ".jpg" in x]
-    xmls = [x.replace(".jpg", ".xml") for x in images]
+    xmls = []
+    for img in images:
+        xml_name = img.replace(".jpg", ".xml")
+        if os.path.isfile(xml_name):
+            xmls.append(xml_name)
 
-    print("copying images")
+    logging.info(f"Number of images found: {len(images)}")
+    logging.info(f"Number of xmls found: {len(xmls)}")
+    if len(images) == 0:
+        logging.critical("No Images found!")
+        sys.exit(1)
+    if len(xmls) == 0:
+        logging.critical("No XMLs found!")
+        sys.exit(1)
+    if len(xmls) != len(images):
+        logging.critical("Number of Images and XMLs Do not Match!")
+        sys.exit(1)
+
+    logging.info("Found images and xmls.")
+
+    logging.info("Creating empty VOC directory skeleton")
+    create_blank_voc_dir_str(dest_path)
+
+    print(" Copying Images ".center(80, "*"))
 
     for img_path in tqdm(images):
         src = img_path
         dest = os.path.join(dest_jpeg_path, img_path.split("/")[-1])
         shutil.copyfile(src, dest)
 
-    print("copying xmls")
+    print(" Copying XMLs ".center(80, "*"))
 
+    missed_xmls = 0
     for xml_path in tqdm(xmls):
         src = xml_path
-        dest = os.path.join(dest_annotation_path, xml_path.split("/")[-1])
-        shutil.copyfile(src, dest)
+        if not os.path.exists(src):
+            logging.critical(f"Missing xml: ({src.split('/')[-1]})")
+            missed_xmls += 1
+        else:
+            dest = os.path.join(dest_annotation_path, xml_path.split("/")[-1])
+            shutil.copyfile(src, dest)
+    if missed_xmls > 0:
+        logging.critical(f"Number of missing xmls: {missed_xmls}")
+        sys.exit(1)
+    else:
+        total_images = len(images)
+        test_ratio = int(ratio * total_images)
 
-    total_images = len(images)
-    test_ratio = int(ratio * total_images)
+        img_names = [x.split("/")[-1].split(".")[0] for x in images]
+        img_names = np.random.permutation(img_names)
+        test_images = img_names[:test_ratio]
+        train_images = img_names[test_ratio:]
+        print(" Working on Train Test Split ".center(80, "*"))
 
-    img_names = [x.split("/")[-1].split(".")[0] for x in images]
-    img_names = np.random.permutation(img_names)
-    test_images = img_names[:test_ratio]
-    train_images = img_names[test_ratio:]
-    print("Working on Train Test Split")
+        with open(os.path.join(dest_imgset_path, "trainval.txt"), "w") as tvf:
+            for img_name in train_images:
+                tvf.write(img_name + "\n")
 
-    with open(os.path.join(dest_imgset_path, "trainval.txt"), "w") as tvf:
-        for img_name in train_images:
-            tvf.write(img_name + "\n")
-
-    with open(os.path.join(dest_imgset_path, "test.txt"), "w") as tf:
-        for img_name in test_images:
-            tf.write(img_name + "\n")
+        with open(os.path.join(dest_imgset_path, "test.txt"), "w") as tf:
+            for img_name in test_images:
+                tf.write(img_name + "\n")
+        print(" Dataset Creation Complete ".center(80, "*"))
 
 
 def main() -> int:
@@ -81,7 +115,6 @@ def main() -> int:
         "-s",
         "--source_path",
         type=str,
-        default="./",
         help="Provide the full path where all the images and xmls are located. Please make sure the img_names are UNIQUE",
     )
